Fixed :multipart clobbering Authorization, User-Agent headers
Supplying the :multipart option caused a conditional branch
in coerce-req to be followed which failed to retain the work
done by prepare-request-headers earlier in the function.

The association of the new Content-Type value was also
simplified by using assoc-in. Added a test for the private
coerce-req to expose the issue and verify the fix.

Signed-off-by: Christian Romney <cromney@pointslope.com>