[#259 #227] Reset decoder after starting the close handshake for the websocket server (@venantius)
In org.http.server.HttpServer.decodeWs(...), when we receive a CloseFrame, we
don't reset the decoder. This eventually causes the server to run out of memory
as the server just sends CLOSE infinitely.

This commit adds a call to reset the decoder so that we don't get stuck in the
infinite close loop.

Resolves #227