[cleanup] Adding type hints to tests
This project has *warn-on-reflection* set to true. This generates a lot
of extra output at test time. This commit adds appropriate type
annotations to the cases where we were getting reflection warnings in
all but one case.

Note that this unfortunately still leaves in place the reflection
warnings for http-kit's dependencies.