Error while processing chunked responses split between CRLF characters
HTTP chunked transfer encoding requires the response to end in CRLF characters. However, it is possible that the CR and LF characters are split into two TCP packets and will not be read into the same byte buffer. CR will be the last byte of one read and LF will be the only content of the next read. The logic in Decoder.readEmptyLine assumes that decoder reaches ALL_READ if CR is the last character in the buffer. Later when LF arrives in the next packet it will be ignored.

This implies that when HTTPClient.doRead() invokes Decoder.decode(), it will detect that the decoder reached ALL_READ state twice. First on reading the CR character and then on the LF character. Then HTTPClient.doRead() creates two PersistentConn objects for the same SelectionKey and adds them to the keepalives queue. So we end up with two keep-alive connections both using the same channel.

This is a recipe for disaster. If two subsequent concurrent requests manage to attach to these two keep-alive connections, they will send their requests on the same channel. But both responses will be processed by the handler of the last request that attached to the SelectionKey. Thus one of those two requests will get the response of the other request.