Graceful Shutdown
Shutdown more gracefully in the presence of HTTP keep alives.

A two phase approach:

- keep track of the currently kept-alive connections in a set, so that
we can close them when we are shutting down.
- Set a boolean flag true when we are shutting down, so that we can use
it to force isKeepAlive to false when shutting down.

Bonus Features:

Change the error message in RingHandler,.java to be correct when the
ExecutorService is being shutdown. This error message happens when the
server is overloaded _and_ when it is being shutdown.

Add logging to the empty `catch` blocks in HttpServer.java