[fix] [Server] [#375] [EXPERIMENTAL] NB prevent unintentional re-use of channels (@osbert)
It can be dangerous to re-use channels for new HTTP requests.

A user might hold on to a channel, and reasonably expect it to *stay closed* after
use. By re-using channels, we risk a user unintentionally using a channel held from
req1 that's since been RE-OPENED for req2 (e.g. a completely unrelated request).

This is more than theoretical- the issue affected Sente,
  Ref. https://github.com/ptaoussanis/sente/issues/260

Big thanks to @osbert for identifying this risk, and for the proposed test and fix.

Commit marked as EXPERIMENTAL since possible performance implications of this change
have not yet been measured, and alternative solutions not yet exhaustively considered.

Still, since the behaviour prior to this commit is dangerous - it seems reasonable
to adopt the present fix as preliminary until further work can be done to either:

  1. Establish that this approach has limited or acceptable downsides, or
  2. A better alternative can be found