[fix] [Client] [#464] Retain dynamic client on client redirects
In hindsight, it's probably better to do this with a simple :client
arg in the recursive call rather than using (bound-fn* request) since
there's currently no other relevant dynamic vars.

And note that if the user wants their callback fn to retain bindings,
it'd be better for them to indicate that directly with their own
(bound-fn* callback).