[fix] [client] [#528] Possible fix for broken `insecure?` client option
From my understanding, it looks like the purpose of this ~undocumented
option was to allow the client to connect to untrusted SSL certificates.

I'm not familiar with how this would normally be accomplished, am short
on time, and no one else has volunteered to tackle the issue - so am just
trying something simple here which is to revert to the legacy (non-SNI)
client when `:insecure?` is true.

This fixes the pre-existing tests, and allows the client to be used
with at least basic self-signed certificates, etc.

Input and/or PRs welcome if anyone has a suggestion for a better way
of doing this.