Fix #578
The change from #375 was preventing the close handler from firing.
This was due to creating new AsyncChannel which did not have a close
handler set.

To fix this we only create a new channel if the previous one was
closed. This prevents the issue of closed channels being reopened and
reused described in #375. But, still ensures the close handler for a
channel is called correctly. AsyncChannel channel new
AsyncChannel(key, this);